+++ miopengemm/include/miopengemm/randomutil.hpp
@@ -17,13 +17,13 @@ class RandomUtil
   private:
   std::random_device                      rd;
   std::default_random_engine              gen;
-  std::uniform_int_distribution<unsigned> unidis;
+  std::uniform_int_distribution<size_t> unidis;
 
   public:
   RandomUtil();
-  unsigned get_from_range(unsigned upper);
+  size_t get_from_range(size_t upper);
   template <typename T>
-  void shuffle(unsigned start_index, unsigned end_index, T& t)
+  void shuffle(size_t start_index, size_t end_index, T& t)
   {
     if (end_index > t.size() || start_index > end_index)
     {
