+++ miopengemm/include/miopengemm/redirection.hpp
@@ -14,12 +14,12 @@ void redirect(bool&          isColMajor,
               bool&          tA,
               bool&          tB,
               bool&          tC,
-              unsigned&      m,
-              unsigned&      n,
-              unsigned&      lda,
-              unsigned&      ldb,
-              unsigned&      a_offset,
-              unsigned&      b_offset,
+              size_t&      m,
+              size_t&      n,
+              size_t&      lda,
+              size_t&      ldb,
+              size_t&      a_offset,
+              size_t&      b_offset,
               const TFloat*& a,
               const TFloat*& b);
 
@@ -27,12 +27,12 @@ void redirect(bool&        isColMajor,
               bool&        tA,
               bool&        tB,
               bool&        tC,
-              unsigned&    m,
-              unsigned&    n,
+              size_t&    m,
+              size_t&    n,
               std::string& a,
               std::string& b);
 
-void confirm_redirection(bool isColMajor, bool tA, bool tB, unsigned m, unsigned n);
+void confirm_redirection(bool isColMajor, bool tA, bool tB, size_t m, size_t n);
 
 }
 }
